+++ always-allow-admin-comments.php
@@ -30,10 +30,11 @@
  * - Add filter for support post types
  * - Add support for new block editor (aka Gutenberg) (once it becomes possible to alter
  *   the contents of the discussion panel)
+ * - Check if post type supports comments before adding checkbox to Discussion box?
  */
 
 /*
-	Copyright (c) 2013-2019 by Scott Reilly (aka coffee2code)
+	Copyright (c) 2013-2020 by Scott Reilly (aka coffee2code)
 
 	This program is free software; you can redistribute it and/or
 	modify it under the terms of the GNU General Public License