+++ always-allow-admin-comments.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name: Always Allow Admin Comments
- * Version:     1.2.1
+ * Version:     1.2.2
  * Plugin URI:  http://coffee2code.com/wp-plugins/admin-can-always-comment/
  * Author:      Scott Reilly
  * Author URI:  http://coffee2code.com/
@@ -18,7 +18,7 @@
  *
  * @package Always_Allow_Admin_Comments
  * @author  Scott Reilly
- * @version 1.2.1
+ * @version 1.2.2
  */
 
 /*
@@ -91,7 +91,7 @@ class c2c_AlwaysAllowAdminComments {
 	 * @since 1.0
 	 */
 	public static function version() {
-		return '1.2.1';
+		return '1.2.2';
 	}
 
 	/**