+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## _(in-progress)_
 * Change: Note compatibility through WP 5.3+
+* Change: Update copyright date (2020)
 
 ## 1.2.1 _(2019-06-18)_
 * Change: Update unit test install script and bootstrap to use latest WP unit test repo