+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## _(in-progress)_
-* New: Unit tests: Add test to verify plugin hooks `plugins_loaded` action to initialize itself
+* Unit tests:
+    * New: Add test to verify plugin hooks `plugins_loaded` action to initialize itself
+    * New: Add unit tests for function `can_show_ui()`
 * Change: Note compatibility through WP 5.3+
 * Change: Update copyright date (2020)