+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## _(in-progress)_
+## 1.2.2 _(2019-12-28)_
 * Unit tests:
     * New: Add test to verify plugin hooks `plugins_loaded` action to initialize itself
     * New: Add unit tests for functions `can_show_ui()`, `comments_open_for_admin()`